Create ManifestGenerator base class
