+++ README.md
@@ -9,36 +9,42 @@ Provides experimental support for the [Asset Manifest RFC](https://github.com/em
 
 Ember Asset Loader does three primary things:
 
-1. Allows you to generate an Asset Manifest,
+1. Provides a base class to easily generate an Asset Manifest,
 2. Provides an Ember service to use the generated manifest at runtime, and
 3. Initializes the above service with the above generated manifest.
 
 ### Generating an Asset Manifest
 
-Generating a manifest is straightforward. Simply import the `generateAssetManifest` from `ember-asset-loader/lib/generate-asset-manifest` during `postprocessTree`:
+You can generate an Asset Manifest by creating either a standalone or in-repo addon which extends from the
+`ManifestGenerator` base class:
 
 ```js
-postprocessTree: function(type, tree) {
-  if (type === 'all') {
-    var generateAssetManifest = require('ember-asset-loader/lib/generate-asset-manifest');
-    return generateAssetManifest(tree, {
-      supportedTypes: [ 'js', 'css' ],
-      bundlesLocation: 'engine-dist'
-    });
+var ManifestGenerator = require('ember-asset-loader/lib/manifest-generator');
+module.exports = ManifestGenerator.extend({
+  name: 'asset-generator-addon',
+  manifestOptions: {
+    bundlesLocation: 'engines-dist',
+    supportedTypes: [ 'js', 'css' ]
   }
-
-  return tree;
-}
+});
 ```
 
-The `generateAssetManifest` function will generate an asset manifest and merge it back into the provided tree. It can
-also take in an option hash with the following supported fields:
+The `ManifestGenerator` will generate an asset manifest and merge it into your build tree during post-processing. It
+generates the manifest according to the options specified in `manifestOptions`:
 
-* The `bundlesLocation` option is a string that specifies which directory in the tree contains the bundles to be placed
-into the asset manifest. This defaults `bundles`.
+* The `bundlesLocation` option is a string that specifies which directory in the build tree contains the bundles to be
+placed into the asset manifest. This defaults `bundles`.
 
-* The `supportedTypes` option is an array that specifies which types of files should be included into the bundles for the
-asset manifest. This defaults to `[ 'js', 'css' ]`.
+* The `supportedTypes` option is an array that specifies which types of files should be included into the bundles for
+the asset manifest. This defaults to `[ 'js', 'css' ]`.
+
+_Note: This class provides default `contentFor`, `postprocessTree`, and `postBuild` hooks so be sure that you call
+`_super` if you override one of those methods._
+
+### Why isn't a manifest generated by default?
+
+No manifest is generated by default since there is no convention for bundling assets within Ember currently. Thus, to
+prevent introducing unintuitive or conflicting behavior, we provide no default generation.
 
 If no manifest is generated, you'll get a warning at build time to ensure that you understand no manifest has been
 generated and thus you'll have to provide a manifest manually in order to use the Asset Loader Service. This warning can
@@ -51,10 +57,6 @@ var app = new EmberApp(defaults, {
   }
 });
 ```
-
-### Why isn't a manifest generated by default?
-
-No manifest is generated by default since there is no convention for bundling assets within Ember currently. Thus, to prevent introducing unintuitive or conflicting behavior, we provide no default generation.
 
 ## Installation
